Alteração do tipo de dados na tabelaEntrega no campo quantidade, tabelaCliente nos campos idade e contacto

+++ app/src/main/java/pt/ipg/app_entregas/EditarClienteFragment.kt
@@ -113,7 +113,7 @@ class EditarClienteFragment : Fragment() {
         }
     }
         private fun alteraCliente(nome: String, contacto: String, idade: String, morada: String): Boolean {
-            val cliente = Cliente(nome, contacto, idade, morada)
+            val cliente = Cliente(nome, contacto.toInt(), idade.toInt(), morada)
 
             val enderecoClientes = Uri.withAppendedPath(
                 ContentProviderEntregas.ENDERECO_CLIENTE,
@@ -131,7 +131,7 @@ class EditarClienteFragment : Fragment() {
         }
 
         private fun insereCliente(nome: String, contacto: String, idade: String, morada: String): Boolean {
-            val cliente = Cliente(nome, contacto, idade, morada)
+            val cliente = Cliente(nome, contacto.toInt(), idade.toInt(), morada)
 
             val enderecoClienteInserido = requireActivity().contentResolver.insert(
                 ContentProviderEntregas.ENDERECO_CLIENTE,
